refactor: getLottoNumbers 메서드 Purcahse 클래스로 이동
