+++ __tests__/LottoTest.js
@@ -1,5 +1,4 @@
 /* eslint-disable max-lines-per-function */
-const App = require('../src/App');
 const Lotto = require('../src/Lotto');
 
 describe('로또 클래스 테스트', () => {
@@ -17,33 +16,4 @@ describe('로또 클래스 테스트', () => {
   });
 
   // 아래에 추가 테스트 작성 가능
-});
-
-const removeDuplicatedNumbers = (numbers) => new Set(numbers);
-
-describe('로또 번호 생성 테스트', () => {
-  const { getNumbers } = new App();
-
-  test('6개의 번호가 있다. ', () => {
-    const numbers = getNumbers();
-
-    expect(numbers).toHaveLength(6);
-  });
-
-  test('중복된 수가 존재하지 않는다.', () => {
-    const numbers = getNumbers();
-    const result = removeDuplicatedNumbers(numbers);
-
-    expect(result.size).toEqual(numbers.length);
-  });
-
-  test('1~45범위의 수 이어야 한다.', () => {
-    const numbers = getNumbers();
-
-    numbers.forEach((number) => {
-      expect(number).toBeGreaterThanOrEqual(1);
-      expect(number).toBeLessThanOrEqual(45);
-      expect(typeof number).toEqual('number');
-    });
-  });
 });
