+++ src/controller/Purchase.js
@@ -1,0 +1,9 @@
+const { Random } = require('@woowacourse/mission-utils');
+
+class Purchase {
+  getLottoNumbers() {
+    return Random.pickUniqueNumbersInRange(1, 45, 6);
+  }
+}
+
+module.exports = Purchase;
